add smart contract to store information about the patients
This smart contract stores the following information for each patient:
- first name
- last name
- date.
A mapping is created to store these information using as its index the CPF (only numbers, represented as an uint256).

The brazilian CPF consists of 11 digits, e.g., 000.000.000-00 and all non-numeric characters are stripped from the input before using it as the argument of the mapping.