add contract to store clinical data
Add a smart contract to store information about procedures and vaccines for a patient.

The key to the mapping containing the data is the brazilian CPF of the patient. The CPF is first stripped out of any non-numeric characters and then the 11-digit number is used as an uitn256 argument for the mapping.