add test scripts to include an encrypted procedure
Add test scripts to include a new medical procedure as encrypted data.

The procedure data is originally a JSON struct like in the unencrypted version. See the ClinicalData.sol for the definition of the struct.
Then the JSON is stringified and encrypted before it is stored in the blockchain.

The encryption and decryption is performed by the Crypto-js node module.