create deploy script
Deploy all the contracts:
- npx hardhat run scripts/deploy.js --network localhost

The deployment script creates/updates the addresses of the contracts and its respective ABI files on the front-end folder.
The addresses are stored in the config.json file on the src folder of the front-end part.
The ABIs are located in a folder called abis also in the src folder of the front-end part.