retrieve employee data from the blockchain
Retrieve employee info from the blockchain and update the input fields with the respective information about the connected employee.

It is possible to edit the values and, by doing so, it is possible to update the information about the employee.
Changing the address to a new one means to add a new employee.