add function to include a vaccine as encrypted data
The addEncryptedVaccines function is defined in the ClinicalData.sol smart contract. See the smart contract source code for the implementation of that function and for the declaration of the nested mapping for storing the encrypted data.

- Parse the input fields into a JSON object (same data struct as in the plain text version defined in the ClinicalData.sol smart contract)
- Encrypt the stringified the JSON object using the public key.

The smart contract function uses the CPF defined in the Patient section to be the key of the nested mapping of the vaccine info. So it is passed as another argument to the addEncryptedVaccines function.