add functions to retrieve the personal info and the procedures for a patient
On changing the  of the patient, load the personal information and load the procedures for the patient.
- display the information in the React component
- as the encrypted and the plain text procedures are stored in different nested mappings and must be trated diferently, the total number of procedures is the sum of both totals

The function getPatient takes as argument the CPF of the patient and returns a struct with the personal information of the patient.

To load the (encrypted/plain text) procedures, first the total number of procedures is loaded, and then the respective nested mapping is iterated upon and the data from the procedure is retrieved from the blockchain.