add function to retrieve the data of the vaccines
On changing the  of the patient, load the personal information and load the vaccines for the patient.
- display the information in the React component
- as the encrypted and the plain text vaccines are stored in different nested mappings and must be trated diferently, the total number of vaccines is the sum of both totals

Add an interface for the retrieved vaccines to be used in the hooks type. The interface is defined by the struct of the data stored and can be found in the ClinicalData.sol smart contract.

The function getPatient takes as argument the CPF of the patient and returns a struct with the personal information of the patient.

To load the (encrypted/plain text) vaccines, first the total number of vaccines is loaded, and then the respective nested mapping is iterated upon and the data from the vaccine is retrieved from the blockchain.